Add tests for timer persistence and recovery
Add a dedicated persistence test suite and shared test utilities, and refactor existing timer-state tests.

- Add tests/unit-js/timer-persistence.test.mjs: exercises settings and timer-state persistence, JSON corruption fallbacks, paused vs running recovery, and timestamp-based hydration.
- Add tests/unit-js/timer-state.test-utils.mjs: VM-based loader for src/main/resources/static/js/timer-state.js, an in-memory localStorage mock, helpers to start a pomodoro and to mock Date.now().
- Update tests/unit-js/timer-state.test.mjs to import/use the new utilities and remove duplicated persistence tests.
- Update tests/unit-js/README.md to document the executable test suites.

This refactors persistence-related tests out of the main state tests, centralizes runtime helpers for isolation, and improves test reuse and clarity.
